Fix SoL and KVM refresh bug
Clicking the refresh button in the app header on the SoL and
KVM pages would fail to reload because the wrong element reference
is passed to the contructor after the reload. Angular recommends
using directives for any DOM manipulation.

- Use the $element ref availble in serial-console.js directive to
  select DOM element to open the terminal
- Switch kvm controller to directive to be able to use element ref
  in the link function to select DOM element to pass as target
