+++ app/server-control/controllers/kvm-controller.html
@@ -1,5 +1,1 @@
-<div id="noVNC_container">
-    <div id="noVNC_status_bar">
-        <div id="noVNC_left_dummy_elem"></div>
-    </div>
-</div>
+<kvm-console></kvm-console>
